# docs/docs/openapi.yml
openapi: 3.0.0

info:
  version: 0.0.1
  title: Custom API for Schulmanager

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    teacher:
      type: object
      properties:
        id:
          type: integer
        firstname:
          type: string
          example: John
        lastname:
          type: string
          example: Doe
        abbreviation:
          type: string
          example: Do
    course:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
          example: Computer Science
        abbreviation:
          type: string
          example: CS
        teacher:
          $ref: "#/components/schemas/teacher"
        color:
          type: string

paths:
  /session:
    description: manage session (logging in and out etc.)
    get:
      security:
        - bearerAuth: [ ]
      description: get info about session
      responses:
        200:
          description: successfully fetched login info
          content:
            application/json:
              schema:
                type: object
                properties:
                  email:
                    type: string
                    example: john.doe@example.com
                  id:
                    type: integer
                  firstname:
                    type: string
                    example: John
                  lastname:
                    type: string
                    example: Doe
                  associatedStudent:
                    type: object
                    properties:
                      id:
                        type: integer
                      firstname:
                        type: string
                        example: John
                      lastname:
                        type: string
                        example: Doe

    post:
      description: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                  required: true
                password:
                  type: string
                  required: true
      responses:
        201:
          description: Logging in was successful
          content:
            text/plain:
              schema:
                type: string
                description: Bearer JWT

  /schedule:
    get:
      security:
        - bearerAuth: [ ]
      parameters:
        - name: start
          schema:
            type: string
            example: 2021-11-08
          in: query
        - name: end
          schema:
            type: string
            example: 2021-11-14
          in: query
      responses:
        200:
          description: successfully fetched schedule
          content:
            application/json:
              schema:
                type: object
                properties:
                  hours:
                    type: array
                    items:
                      type: object
                      properties:
                        date:
                          type: string
                          example: 2021-11-10
                          description: date
                        hour:
                          type: integer
                          description: hour id (see GET /schedule/hours)
                        room:
                          type: string
                          example: G053
                        cancelled:
                          type: boolean
                          example: false
                        substitue:
                          type: boolean
                          example: true
                        substituted:
                          type: object
                          description: the original lesson
                          properties:
                            room:
                              type: string
                              example: G036
                            subject:
                              type: object
                              properties:
                                abbreviation:
                                  type: string
                                  example: E
                                name:
                                  type: string
                                  example: English
                                label:
                                  type: string
                                  example: E9
                            teachers:
                              type: array
                              items:
                                type: object
                                properties:
                                  abbreviation:
                                    type: string
                                    example: D2
                                  firstname:
                                    type: string
                                    example: John2
                                  lastname:
                                    type: string
                                    example: Doe2
                        subject:
                          type: object
                          properties:
                            abbreviation:
                              type: string
                              example: CS
                            name:
                              type: string
                              example: Computer Science
                            label:
                              type: string
                              example: CS9

  /schedule/hours:
    get:
      security:
        - bearerAuth: [ ]
      responses:
        200:
          description: fetched hours
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: hour id
                    number:
                      type: string
                      description: number/display name
                    from:
                      type: string
                      description: time of the begin (format HH:mm:ss)
                    until:
                      type: string
                      description: time of the end (format HH:mm:ss)

  /chat:
    get:
      description: fetch all threads
      responses:
        200:
          description: successfully fetched threads
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                      description: id used for referencing in api (/chat/:id)
                    threadId:
                      type: integer
                    sender:
                      type: string
                      example: John Doe
                    recipients:
                      type: string
                      example: Jane Doe
                    privateChat:
                      type: boolean
                    allowsAnswers:
                      type: boolean
                    unread:
                      type: integer
                      description: number of unread messages
                    lastMessage:
                      type: string
                      description: ISO timestamp of last message
      security:
        - bearerAuth: [ ]
    post:
      description: create new thread
      responses:
        201:
          description: successfully created new thread
          content:
            text/plain:
              schema:
                type: integer
                description: id of new thread
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                recipientId:
                  type: integer
                threadName:
                  type: string
                text:
                  type: string
                  description: content of first message
                files:
                  type: array
                  description: attachments of first message
                  items:
                    type: string
                    format: binary
  /chat/{id}:
    get:
      description: fetch messages in thread
      security:
        - bearerAuth: [ ]
      responses:
        200:
          description: successfully fetched messages
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    text:
                      type: string
                    sentTimestamp:
                      type: string
                      description: ISO timestamp
                    sender:
                      type: object
                      properties:
                        firstname:
                          type: string
                        lastname:
                          type: string
                    attachments:
                      type: array
                      items:
                        type: object
                        properties:
                          url:
                            type: string
                            description: download URL
                          mime:
                            type: string
                            description: file mime type
                          filename:
                            type: string
      parameters:
        - in: path
          name: id
          schema:
            type: integer

    post:
      description: send message
      security:
        - bearerAuth: [ ]
      responses:
        201:
          description: successfully sent message
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                text:
                  type: string

  /chat/users:
    get:
      responses:
        200:
          description: successfully fetched chat users
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: integer
                    name:
                      type: string
                    subtext:
                      type: string
                      description: for example which subjects they teach
      description: get all users who can be sent messages